+++ spec/requests/as_account/channel_spec.rb
@@ -215,8 +215,8 @@ describe Yt::Channel, :device_app do
       expect{channel.card_teaser_clicks}.not_to raise_error
       expect{channel.card_teaser_click_rate}.not_to raise_error
       expect{channel.viewer_percentage}.not_to raise_error
-      expect{channel.earnings}.to raise_error Yt::Errors::Unauthorized
-      expect{channel.impressions}.to raise_error Yt::Errors::Unauthorized
+      expect{channel.estimated_revenue}.to raise_error Yt::Errors::Unauthorized
+      expect{channel.ad_impressions}.to raise_error Yt::Errors::Unauthorized
       expect{channel.monetized_playbacks}.to raise_error Yt::Errors::Unauthorized
       expect{channel.playback_based_cpm}.to raise_error Yt::Errors::Unauthorized
     end